+++ package.json
@@ -9,10 +9,8 @@
   "author": "",
   "license": "ISC",
   "description": "",
-  "dependencies": {
-    "tailwindcss": "^4.0.4"
-  },
   "devDependencies": {
-    "daisyui": "^4.12.23"
+    "daisyui": "^4.12.23",
+    "tailwindcss": "^4.0.6"
   }
 }
