+++ tailwind.config.js
@@ -1,8 +1,12 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
-  content: ["./templates/*.html"],
+  content: ["./templates/**/*.html", "./static/src/**/*.js"],
   theme: {
-    extend: {},
+    extend: {
+      fontFamily: {
+        knewave: ["Knewave", "sans-serif"],
+      },
+    },
   },
-  plugins: [require("daisyui")],
+  plugins: [],
 };
